change filter service folder #3376
