+++ src/Controller/Back/PartnerController.php
@@ -23,11 +23,11 @@ use App\Repository\JobEventRepository;
 use App\Repository\PartnerRepository;
 use App\Repository\UserRepository;
 use App\Security\Voter\PartnerVoter;
+use App\Service\ListFilters\SearchPartner;
 use App\Service\Mailer\NotificationMail;
 use App\Service\Mailer\NotificationMailerRegistry;
 use App\Service\Mailer\NotificationMailerType;
 use App\Service\Sanitizer;
-use App\Service\SearchPartner;
 use App\Service\Signalement\VisiteNotifier;
 use App\Validator\EmailFormatValidator;
 use Doctrine\ORM\EntityManagerInterface;
