+++ src/Controller/Back/TagController.php
@@ -10,7 +10,7 @@ use App\Form\SearchTagType;
 use App\Manager\TagManager;
 use App\Repository\TagRepository;
 use App\Service\FormHelper;
-use App\Service\SearchTag;
+use App\Service\ListFilters\SearchTag;
 use App\Service\Signalement\SearchFilterOptionDataProvider;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
