+++ src/Form/SearchAutoAffectationRuleType.php
@@ -3,7 +3,7 @@
 namespace App\Form;
 
 use App\Entity\Territory;
-use App\Service\SearchAutoAffectationRule;
+use App\Service\ListFilters\SearchAutoAffectationRule;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
