+++ src/Form/SearchPartnerType.php
@@ -5,7 +5,7 @@ namespace App\Form;
 use App\Entity\Enum\PartnerType;
 use App\Entity\Territory;
 use App\Entity\User;
-use App\Service\SearchPartner;
+use App\Service\ListFilters\SearchPartner;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\Form\AbstractType;
