+++ src/Form/SearchTerritoryType.php
@@ -2,7 +2,7 @@
 
 namespace App\Form;
 
-use App\Service\SearchTerritory;
+use App\Service\ListFilters\SearchTerritory;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
