+++ src/Repository/PartnerRepository.php
@@ -8,8 +8,8 @@ use App\Entity\Enum\Qualification;
 use App\Entity\Partner;
 use App\Entity\Signalement;
 use App\Entity\Territory;
-use App\Service\SearchArchivedPartner;
-use App\Service\SearchPartner;
+use App\Service\ListFilters\SearchArchivedPartner;
+use App\Service\ListFilters\SearchPartner;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\DBAL\Exception;
 use Doctrine\ORM\Query\QueryException;
