+++ src/Repository/UserRepository.php
@@ -5,8 +5,8 @@ namespace App\Repository;
 use App\Dto\CountUser;
 use App\Entity\Territory;
 use App\Entity\User;
-use App\Service\SearchArchivedAccount;
-use App\Service\SearchUser;
+use App\Service\ListFilters\SearchArchivedAccount;
+use App\Service\ListFilters\SearchUser;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\ORM\NonUniqueResultException;
 use Doctrine\ORM\QueryBuilder;
