+++ src/Service/ListFilters/SearchTag.php
@@ -1,21 +1,20 @@
 <?php
 
-namespace App\Service;
+namespace App\Service\ListFilters;
 
-use App\Entity\Enum\ZoneType;
 use App\Entity\Territory;
 use App\Entity\User;
 use App\Service\Behaviour\SearchQueryTrait;
 
-class SearchZone
+class SearchTag
 {
     use SearchQueryTrait {
         getUrlParams as getUrlParamsBase;
     }
+
     private User $user;
-    private ?string $queryName = null;
+    private ?string $queryTag = null;
     private ?Territory $territory = null;
-    private ?ZoneType $type = null;
 
     public function __construct(User $user)
     {
@@ -30,14 +29,14 @@ class SearchZone
         return $this->user;
     }
 
-    public function getQueryName(): ?string
+    public function getQueryTag(): ?string
     {
-        return $this->queryName;
+        return $this->queryTag;
     }
 
-    public function setQueryName(?string $queryName): void
+    public function setQueryTag(?string $queryTag): void
     {
-        $this->queryName = $queryName;
+        $this->queryTag = $queryTag;
     }
 
     public function getTerritory(): ?Territory
@@ -50,16 +49,6 @@ class SearchZone
         $this->territory = $territory;
     }
 
-    public function getType(): ?ZoneType
-    {
-        return $this->type;
-    }
-
-    public function setType(?ZoneType $type): void
-    {
-        $this->type = $type;
-    }
-
     public function getUrlParams(): array
     {
         $params = $this->getUrlParamsBase();
@@ -68,5 +57,18 @@ class SearchZone
         }
 
         return $params;
+    }
+
+    public function getFiltersToText(): array
+    {
+        $filters = [];
+        if ($this->queryTag) {
+            $filters['Recherche'] = $this->queryTag;
+        }
+        if ($this->territory && $this->user->isSuperAdmin()) {
+            $filters['Territoire'] = $this->territory->getZip().' - '.$this->territory->getName();
+        }
+
+        return $filters;
     }
 }
